% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stp.R
\name{stpbias}
\alias{stpbias}
\title{Spatio-Temporal Proximity Index Bias Estimation}
\usage{
stpbias(data = D, x = "long", y = "lat", t = "time", ID = "ID",
  group = "race", cores = 2, small = FALSE, sample_size = 100,
  nreps = 300)
}
\description{
\code{stpbias} calculates the spatio-temporal proximity index for a population in
which each individual is labelled as belonging to one of two groups.
}
\details{
This function requires a set of pairwise distances between all individuals in the population and a variable that contains the group labels. It returns the spatio-temporal proximity index and, optionally, a set of boostrap replicates from which uncertainty may be estimated.
}
