% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stp.R
\name{stpesd}
\alias{stpesd}
\title{Spatio-Temporal Proximity Index Empirical Sampling Distribution}
\usage{
stpesd(data = D, x = "long", y = "lat", t = "time", ID = "ID",
  group = "race", distance_function = "NegExp", cores = 2,
  small = FALSE, sample_size = 100, nreps = 300)
}
\description{
\code{stpesd} calculates the spatio-temporal proximity index for a population in
which each individual is labelled as belonging to one of two groups.
}
\details{
This function requires a set of pairwise distances between all individuals in the population and a variable that contains the group labels. It returns the spatio-temporal proximity index and, optionally, a set of boostrap replicates from which uncertainty may be estimated.
}
